refactor: edit sidebar logic for changing semester
