+++ src/utils/itemUtils.tsx
@@ -46,3 +46,14 @@ export const canPreviewItem = (item: Item): boolean => {
     )
   );
 };
+
+export const formatFileSize = (bytes: number): string => {
+  if (bytes === 0) return "0B";
+
+  const k = 1024;
+  const sizes = ["B", "KB", "MB", "GB", "TB"];
+
+  const i = Math.floor(Math.log(bytes) / Math.log(k));
+
+  return `${parseFloat((bytes / Math.pow(k, i)).toFixed(1))}${sizes[i]}`;
+};
